move the generic postprocessing functions to a separate dir
